finance-ai-project/requirements.txt
fastapi      
uvicorn
requests
pandas
numpy
python-dotenv